refactor(errors): move eisdir
